requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.6
fastapi-cors==0.0.6
google-cloud-storage==2.18.0
google-cloud-discoveryengine==0.11.14
google-api-python-client==2.138.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-firestore==2.8.0
google-cloud-secret-manager==2.16.0
google-generativeai==0.7.2
googleapis-common-protos==1.66.0
googlemaps==4.10.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.62.3
gunicorn==23.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-google-genai==1.0.8
python-dotenv==0.21.0
requests==2.32.3
starlette==0.41.3
urllib3==2.3.0
uvicorn==0.30.3
vertexai==1.60.0